+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Support `Arbitrary` for integer types
   * Support `Arbitrary` for float types
 * Ability to specify boundaries (`greater`, `greater_or_equal`, `less`, `less_or_equal`, `len_char_min`, `len_char_max`) with expressions or named constants.
+* Add `#[inline]` attribute to trivial functions
 
 ### v0.4.0 - 2023-11-21
 * Support of arbitrary inner types with custom sanitizers and validators.